+++ docs/Platforms/FOSS/Docker/Meteor3/apt/clean.sh
@@ -1,0 +1,3 @@
+apt clean
+apt -y autoclean
+apt -y autoremove