+++ docs/Platforms/FOSS/Docker/Meteor3/apt/update.sh
@@ -1,0 +1,4 @@
+apt update
+apt -y dist-upgrade
+snap refresh
+